Avoid FloatingDecimal rounding
By using a untyped literal in the Custom XPath calculations were getting
promoted to float precision.
Since target is a decimal representing currency amount, the current
outputs like "0.32000000000000006" are not appropriate and introduce
overheads using float precision. The tutorial should demonstrate the use
of explicit xs: type casting and XPath functions for rounding, so that
the output is like "0.32".